self-host: split BillingLive — core /account page owns export + deletion (#479)
- New core FountainWeb.Live.AccountLive at /account: data export and
  account deletion, extracted verbatim from BillingLive. Handlers already
  delegated to core Fountain.Exports / Fountain.Accounts.Deletion.
- BillingLive is purely billing and redirects to /account when
  Billing.enabled?() is false — old bookmarks keep working, community
  instances never render a billing surface.
- Sidebar: unconditional Account link (exact-match highlight so it does
  not double-light on /account/*); Billing link renders only when
  billing is enabled.
- Tests move with the UI: account export/deletion tests now mount
  /account; ee BillingLive test asserts the disabled-state redirect.
- 402 upgrade_url "/account/billing" contract untouched: it only fires
  when billing is enabled, where the page still exists.

Completes the BillingLive extraction future-work item in decisions/0010.

Closes #479
