+++ apps/fountain/test/fountain_web/live/account_export_live_test.exs
@@ -17,7 +17,7 @@ defmodule FountainWeb.AccountExportLiveTest do
   test "the card states that secret values are excluded", %{conn: conn} do
     user = insert_verified_user()
 
-    {:ok, _lv, html} = live(login_user(conn, user), ~p"/account/billing")
+    {:ok, _lv, html} = live(login_user(conn, user), ~p"/account")
 
     assert html =~ "Export your data"
     assert html =~ "secret values are deliberately excluded"
@@ -27,7 +27,7 @@ defmodule FountainWeb.AccountExportLiveTest do
   test "requesting an export enqueues the job and shows pending status", %{conn: conn} do
     user = insert_verified_user()
 
-    {:ok, lv, _html} = live(login_user(conn, user), ~p"/account/billing")
+    {:ok, lv, _html} = live(login_user(conn, user), ~p"/account")
 
     html = render_click(lv, "request_export")
     assert html =~ "Export started"
@@ -41,7 +41,7 @@ defmodule FountainWeb.AccountExportLiveTest do
   test "a second request within the hour is refused", %{conn: conn} do
     user = insert_verified_user()
 
-    {:ok, lv, _html} = live(login_user(conn, user), ~p"/account/billing")
+    {:ok, lv, _html} = live(login_user(conn, user), ~p"/account")
 
     render_click(lv, "request_export")
     html = render_click(lv, "request_export")
@@ -55,7 +55,7 @@ defmodule FountainWeb.AccountExportLiveTest do
     {:ok, export} = Exports.request_export(user)
     :ok = perform_job(AccountExport, %{export_id: export.id, user_id: user.id})
 
-    {:ok, _lv, html} = live(login_user(conn, user), ~p"/account/billing")
+    {:ok, _lv, html} = live(login_user(conn, user), ~p"/account")
 
     assert html =~ "Export ready"
     assert html =~ ~p"/account/exports/#{export.id}/download"
@@ -65,7 +65,7 @@ defmodule FountainWeb.AccountExportLiveTest do
   test "the completion broadcast flips pending to ready without a reload", %{conn: conn} do
     user = insert_verified_user()
 
-    {:ok, lv, _html} = live(login_user(conn, user), ~p"/account/billing")
+    {:ok, lv, _html} = live(login_user(conn, user), ~p"/account")
     assert render_click(lv, "request_export") =~ "generating"
 
     [export] = Repo.all(Export)
