+++ ee/lib/fountain_web/live/billing_live.ex
@@ -5,40 +5,34 @@ defmodule FountainWeb.Live.BillingLive do
 
   Accessible to all authenticated users regardless of subscription status
   (including `past_due` and `canceled`) so they can update payment details.
+
+  Purely billing since #479: data export and account deletion live on the
+  core `/account` page. On a billing-disabled instance this page does not
+  exist — old bookmarks redirect to `/account`.
   """
 
   use FountainWeb, :live_view
 
-  alias Fountain.Accounts.Deletion
   alias Fountain.Billing
-  alias Fountain.Exports
 
   @impl true
   def mount(_params, _session, socket) do
-    socket = FountainWeb.Audited.put_client_ip(socket)
-    user = socket.assigns.current_user
-    {period_start, period_end} = current_month_range()
-    usage = Billing.usage_summary(user.id, period_start, period_end)
+    if Billing.enabled?() do
+      user = socket.assigns.current_user
+      {period_start, period_end} = current_month_range()
+      usage = Billing.usage_summary(user.id, period_start, period_end)
 
-    if connected?(socket) do
-      Phoenix.PubSub.subscribe(Fountain.PubSub, Exports.topic(user.id))
+      {:ok,
+       assign(socket,
+         page_title: "Billing",
+         usage: usage,
+         period_start: period_start,
+         period_end: period_end,
+         stripe_url_loading: false
+       )}
+    else
+      {:ok, redirect(socket, to: ~p"/account")}
     end
-
-    {:ok,
-     assign(socket,
-       page_title: "Billing",
-       usage: usage,
-       period_start: period_start,
-       period_end: period_end,
-       stripe_url_loading: false,
-       delete_confirmation: "",
-       deleting: false,
-       export: Exports.latest_export(user.id),
-       # On a billing-disabled instance the subscription card is noise: the
-       # status is meaningless with the gate off, and the Upgrade button can
-       # only ever flash "Unable to reach Stripe" (#335).
-       billing_enabled: Billing.enabled?()
-     )}
   end
 
   @impl true
@@ -87,94 +81,13 @@ defmodule FountainWeb.Live.BillingLive do
   end
 
   @impl true
-  def handle_event("request_export", _params, socket) do
-    user = socket.assigns.current_user
-
-    case Exports.request_export(user, actor: "self", request_ip: socket.assigns[:client_ip]) do
-      {:ok, export} ->
-        {:noreply,
-         socket
-         |> assign(:export, export)
-         |> put_flash(:info, "Export started — the download will appear here when it is ready.")}
-
-      {:error, {:rate_limited, retry_after}} ->
-        minutes = max(div(retry_after + 59, 60), 1)
-
-        {:noreply,
-         put_flash(
-           socket,
-           :error,
-           "You can request one export per hour. Try again in about #{minutes} minute(s)."
-         )}
-
-      {:error, _reason} ->
-        {:noreply, put_flash(socket, :error, "Could not start the export. Please try again.")}
-    end
-  end
-
-  @impl true
-  def handle_event("confirm_delete_input", %{"email" => email}, socket) do
-    {:noreply, assign(socket, :delete_confirmation, email)}
-  end
-
-  # Typing the account's own email is the confirmation. It is not a security
-  # control — the session already proves who this is — it is a speed bump, so
-  # the irreversible button cannot be hit by reflex.
-  @impl true
-  def handle_event("delete_account", %{"email" => email}, socket) do
-    user = socket.assigns.current_user
-
-    cond do
-      email != user.email ->
-        {:noreply, put_flash(socket, :error, "Type your email address exactly to confirm.")}
-
-      socket.assigns.deleting ->
-        {:noreply, socket}
-
-      true ->
-        socket = assign(socket, :deleting, true)
-
-        case Deletion.delete_user(user,
-               actor: "self",
-               request_ip: socket.assigns[:client_ip]
-             ) do
-          {:ok, _summary} ->
-            # Straight out through the controller so the session is dropped;
-            # a LiveView cannot clear the session cookie itself.
-            {:noreply, redirect(socket, to: ~p"/auth/logout?deleted=1")}
-
-          {:error, {:stripe, _reason}} ->
-            {:noreply,
-             socket
-             |> assign(:deleting, false)
-             |> put_flash(
-               :error,
-               "Your subscription could not be cancelled, so nothing was deleted. " <>
-                 "Please try again, or contact support if it keeps failing."
-             )}
-
-          {:error, _reason} ->
-            {:noreply,
-             socket
-             |> assign(:deleting, false)
-             |> put_flash(:error, "Account deletion failed. Nothing was deleted.")}
-        end
-    end
-  end
-
-  @impl true
-  def handle_info({:export_updated, _user_id}, socket) do
-    {:noreply, assign(socket, :export, Exports.latest_export(socket.assigns.current_user.id))}
-  end
-
-  @impl true
   def render(assigns) do
     ~H"""
     <div class="mx-auto max-w-2xl space-y-8 px-4 py-8">
       <h1 class="text-2xl font-semibold">Billing</h1>
 
       <%!-- past_due banner --%>
-      <%= if @billing_enabled and @current_user.subscription_status == "past_due" do %>
+      <%= if @current_user.subscription_status == "past_due" do %>
         <div class="rounded border border-red-300 bg-red-50 px-4 py-3 text-sm text-red-800" role="alert">
           Your subscription requires attention. Update your payment method to
           continue starting conversations.
@@ -183,7 +96,7 @@ defmodule FountainWeb.Live.BillingLive do
 
       <%!-- cancel-at-period-end notice: canceled in the portal, still inside
            the paid period. Access continues until the period ends. --%>
-      <%= if @billing_enabled and canceling_at_period_end?(@current_user) do %>
+      <%= if canceling_at_period_end?(@current_user) do %>
         <div
           class="rounded border border-amber-300 bg-amber-50 px-4 py-3 text-sm text-amber-800"
           role="alert"
@@ -194,19 +107,8 @@ defmodule FountainWeb.Live.BillingLive do
         </div>
       <% end %>
 
-      <%!-- Billing disabled: no status theater, no button that can only fail --%>
-      <%= if not @billing_enabled do %>
-        <div class="rounded-lg border bg-white p-6 shadow-sm">
-          <h2 class="mb-2 text-lg font-medium">Subscription</h2>
-          <p class="text-sm text-gray-600">
-            Billing is disabled on this instance. All features are available
-            without a subscription.
-          </p>
-        </div>
-      <% end %>
-
       <%!-- Subscription status card --%>
-      <div :if={@billing_enabled} class="rounded-lg border bg-white p-6 shadow-sm">
+      <div class="rounded-lg border bg-white p-6 shadow-sm">
         <h2 class="mb-4 text-lg font-medium">Subscription</h2>
         <dl class="space-y-3">
           <div class="flex items-center justify-between">
@@ -313,120 +215,11 @@ defmodule FountainWeb.Live.BillingLive do
           </div>
         </dl>
       </div>
-
-      <%!-- Data export --%>
-      <div class="rounded-lg border bg-white p-6 shadow-sm">
-        <h2 class="mb-1 text-lg font-medium" id="export">Export your data</h2>
-        <p class="mb-2 text-sm text-gray-600">
-          Download a single JSON file containing your agents, environments, vaults,
-          conversations with their full log output, and your audit trail.
-        </p>
-        <p class="mb-4 text-sm text-gray-600">
-          Environment and vault <strong>secret values are deliberately excluded</strong>
-          — only secret names are listed. Secrets were write-only on the way in and
-          stay that way on the way out.
-        </p>
-
-        <%= if @export do %>
-          <div class="mb-4 rounded-md bg-gray-50 p-4 text-sm" id="export-status">
-            <%= case export_state(@export) do %>
-              <% :pending -> %>
-                <span class="text-gray-700">
-                  Export requested <%= Calendar.strftime(@export.inserted_at, "%Y-%m-%d %H:%M UTC") %>
-                  — generating&hellip; The download will appear here when it is ready.
-                </span>
-              <% :ready -> %>
-                <div class="flex items-center justify-between gap-4">
-                  <span class="text-gray-700">
-                    Export ready (<%= format_bytes(@export.byte_size) %>) — link expires
-                    <%= Calendar.strftime(@export.expires_at, "%Y-%m-%d %H:%M UTC") %>.
-                  </span>
-                  <a
-                    href={~p"/account/exports/#{@export.id}/download"}
-                    class="shrink-0 rounded-md bg-indigo-600 px-3 py-1.5 text-sm font-medium text-white hover:bg-indigo-700"
-                  >
-                    Download
-                  </a>
-                </div>
-              <% :expired -> %>
-                <span class="text-gray-500">
-                  Your last export has expired. Request a new one to download your data.
-                </span>
-              <% :failed -> %>
-                <span class="text-red-700">
-                  The last export failed. Please request a new one; contact support if it
-                  keeps failing.
-                </span>
-            <% end %>
-          </div>
-        <% end %>
-
-        <button
-          phx-click="request_export"
-          class="rounded-md border border-gray-300 bg-white px-4 py-2 text-sm font-medium text-gray-700 hover:bg-gray-50"
-        >
-          Request export
-        </button>
-        <p class="mt-2 text-xs text-gray-400">
-          One export per hour. Downloads expire after <%= Exports.ttl_hours() %> hours.
-        </p>
-      </div>
-
-      <%!-- Danger zone --%>
-      <div class="rounded-lg border border-red-200 bg-white p-6 shadow-sm">
-        <h2 class="mb-1 text-lg font-medium text-red-700">Delete account</h2>
-        <p class="mb-4 text-sm text-gray-600">
-          Cancels your subscription, destroys every running sandbox, and permanently
-          deletes your agents, environments, vaults, conversations and stored secrets.
-          Secrets are encrypted with a key held only for your account; deleting the
-          account destroys that key, so they cannot be recovered afterwards by anyone.
-          <strong>This cannot be undone.</strong>
-        </p>
-        <p class="mb-4 text-sm text-gray-600" id="delete-export-nudge">
-          Want a copy of your data first?
-          <a href="#export" class="underline">Request an export above</a>
-          before deleting — nothing can be recovered afterwards.
-        </p>
-
-        <form phx-submit="delete_account" class="space-y-3">
-          <label class="block text-sm text-gray-700">
-            Type <span class="font-mono font-semibold"><%= @current_user.email %></span> to confirm
-            <input
-              type="text"
-              name="email"
-              autocomplete="off"
-              value={@delete_confirmation}
-              phx-change="confirm_delete_input"
-              class="mt-1 block w-full rounded border-gray-300 text-sm shadow-sm"
-            />
-          </label>
-
-          <button
-            type="submit"
-            disabled={@delete_confirmation != @current_user.email or @deleting}
-            class="rounded bg-red-600 px-4 py-2 text-sm font-medium text-white disabled:cursor-not-allowed disabled:bg-gray-300"
-          >
-            <%= if @deleting, do: "Deleting…", else: "Delete my account" %>
-          </button>
-        </form>
-      </div>
     </div>
     """
   end
 
   # ─── Private helpers ───────────────────────────────────────────────────────────
-
-  defp export_state(%{status: "pending"}), do: :pending
-  defp export_state(%{status: "failed"}), do: :failed
-
-  defp export_state(%{status: "completed"} = export) do
-    if Exports.expired?(export), do: :expired, else: :ready
-  end
-
-  defp format_bytes(nil), do: "unknown size"
-  defp format_bytes(n) when n < 1024, do: "#{n} B"
-  defp format_bytes(n) when n < 1024 * 1024, do: "#{Float.round(n / 1024, 1)} KB"
-  defp format_bytes(n), do: "#{Float.round(n / (1024 * 1024), 1)} MB"
 
   defp current_month_range do
     now = DateTime.utc_now()
